+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Helm: Allow Pod `priorityClassName` to be configured ([#752]).
 - Add support for `34.0.0` ([#755]).
+- Add `prometheus.io/path|port|scheme` annotations to metrics services ([#761]).
 
 ### Changed
 
@@ -31,6 +32,7 @@ All notable changes to this project will be documented in this file.
 [#753]: https://github.com/stackabletech/druid-operator/pull/753
 [#755]: https://github.com/stackabletech/druid-operator/pull/755
 [#756]: https://github.com/stackabletech/druid-operator/pull/756
+[#761]: https://github.com/stackabletech/druid-operator/pull/761
 
 ## [25.7.0] - 2025-07-23